hw 18-1 full
